requirements.txt
absl-py==2.2.2
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
grpcio==1.71.0
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
protobuf==6.30.2
pyparsing==3.2.3
pyslurm==24.11.0
python-dateutil==2.9.0.post0
PyWavelets==1.8.0
scikit-image==0.20.0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tifffile==2025.5.10
torch @ file:///home/balannag/pytorch_wheels/torch-2.3.1%2Bcu121-cp310-cp310-linux_x86_64.whl
torchsummary==1.5.1
torchvision @ file:///home/balannag/pytorch_wheels/torchvision-0.18.1%2Bcu121-cp310-cp310-linux_x86_64.whl
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.13.2
Werkzeug==3.1.3
